compiling again part 2

+++ uarchsim/bct.h
@@ -3,7 +3,7 @@
 class bct{
     int* counters;
     uint64_t get_index(uint64_t pc);
-    uint64_t num_entries = 0;
+    int num_entries = 0;
     public:
     bct(uint64_t size);
     uint64_t get_count(uint64_t pc, uint64_t bhr);
